move interpolation and clip code into separate module
